Start work on implementing mulliken charges
