+++ src/JuliaChem.jl
@@ -6,7 +6,6 @@ Base.include(@__MODULE__, "modules/JCModules.jl")
 Base.include(@__MODULE__, "basis/JCBasis.jl")
 Base.include(@__MODULE__, "io/JCInput.jl")
 Base.include(@__MODULE__, "molecule/JCMolecule.jl")
-Base.include(@__MODULE__, "properties/JCProperties.jl")
 Base.include(@__MODULE__, "rhf/JCRHF.jl")
 Base.include(@__MODULE__, "runtime/JCRuntime.jl")
